Update public API and reorganize folder structure
