+++ packages/core/src/modules/oob/messages/index.ts
@@ -1,0 +1,2 @@
+export * from './OutOfBandMessage'
+export * from './HandshakeReuseMessage'
